Remove scheme and host in favor of baseURL
"Overriding" extensions gets a little weird because you don't override them in a typical sense. The Swift compiler orders scope when deciding what implementation to use for a protocol. It checks the type conforming to the protocol, then looks for protocol extensions in that same module, and finally for protocol extensions in imported modules.

This works mostly as one might expect when re-implementing an extension property or method in the main app, but causes ambiguity and compiler errors if you try to unit test the re-implementation. In the unit test module, Swift either won't see the app's re-implementation or it sees both that one and the default one in the Request module. In the former case, your tests will fail because the app's code isn't being chosen. In the latter case, your test won't compile.