Properly handle + in query string values
Although RFC 3986 allows "+" in a URL, if it's set in a query item name or value, we'll assume that it's meant to be a literal plus sign, and not a pre-encoded space. URLComponents makes the opposite assumption, making it impossible to send a literal plus to an API (because, presumably, your API will see the + and decode it to a space).

The fix is to encode names and values ourselves, taking "+" out of the set of allowed characters. Spaces will instead be encoded to %20 and plus signs to %2B.